requirements.txt
openai>=1.0.0
anthropic>=0.18.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
streamlit>=1.35.0
python-docx>=1.1.2
reportlab>=4.2.0
Pillow>=10.0.0
pywebview>=5.0
